% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{cases}
\alias{cases}
\title{Test Cases Data}
\format{An object of class \code{data.frame} with 10 rows and 9 columns.}
\usage{
data(cases)
}
\description{
10 test cases for the Diagnostics project.
The variables here are
Pn (Pneumonia), Te (Temperature), VTB (Visited a TB location), TB (Tuberculosis), Sm (Smoker), LC (Lung Cancer), Br (Bronchitis),
XR (X-Ray result), Dy (Dyspnea, or shortness of breath). The Pn, TB, LC and Br variables are NA (unknown) in all cases.
}
\examples{
data(cases)
}
\keyword{datasets}
