% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_assumptions.R
\name{check_assumptions}
\alias{check_assumptions}
\title{Check regression assumptions}
\usage{
check_assumptions(data, response, predictors, model_type = "lm")
}
\arguments{
\item{data}{A data frame.}

\item{response}{Unquoted name of the response variable.}

\item{predictors}{A call to c() with unquoted variable names.}

\item{model_type}{Either "lm" (default) or "glm" for logistic regression.}
}
\value{
A named list of diagnostic plots.
}
\description{
Check regression assumptions
}
\examples{
check_assumptions(mtcars, mpg, c(wt, hp, qsec))
check_assumptions(mtcars, am, c(wt, hp), model_type = "glm")
}
