% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_homoscedasticity.R
\name{check_homoscedasticity}
\alias{check_homoscedasticity}
\title{Check homoscedasticity with residuals vs. fitted plot}
\usage{
check_homoscedasticity(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Unquoted name of the response variable.}

\item{predictors}{A call to c() with unquoted variable names.}
}
\value{
A ggplot2 plot of residuals vs. fitted values.
}
\description{
Check homoscedasticity with residuals vs. fitted plot
}
\examples{
check_homoscedasticity(mtcars, mpg, c(wt, hp, qsec))
}
