% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_linearity.R
\name{check_linearity}
\alias{check_linearity}
\title{Check linearity with added variable plots}
\usage{
check_linearity(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Unquoted name of the response variable.}

\item{predictors}{A call to c() with unquoted variable names.}
}
\value{
A list of ggplot2 plots (one per predictor).
}
\description{
Check linearity with added variable plots
}
\examples{
check_linearity(mtcars, mpg, c(wt, hp, qsec))
}
