% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_log_odds_linearity.R
\name{check_log_odds_linearity}
\alias{check_log_odds_linearity}
\title{Check log-odds linearity (logistic regression)}
\usage{
check_log_odds_linearity(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Unquoted binary response variable (0/1).}

\item{predictors}{A call to c() with unquoted variable names.}
}
\value{
A named list of ggplot2 plots (one per predictor).
}
\description{
Check log-odds linearity (logistic regression)
}
\examples{
check_log_odds_linearity(mtcars, am, c(wt, hp))
}
