% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_normality.R
\name{check_normality}
\alias{check_normality}
\title{Check normality of residuals with Q-Q plot}
\usage{
check_normality(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Unquoted name of the response variable.}

\item{predictors}{A call to c() with unquoted variable names.}
}
\value{
A ggplot2 Q-Q plot of residuals.
}
\description{
Check normality of residuals with Q-Q plot
}
\examples{
check_normality(mtcars, mpg, c(wt, hp, qsec))
}
