% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_outliers.R
\name{check_outliers}
\alias{check_outliers}
\title{Check for outliers using Cook's distance}
\usage{
check_outliers(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Unquoted name of the response variable.}

\item{predictors}{A call to c() with unquoted variable names.}
}
\value{
A ggplot2 plot of Cook's distance.
}
\description{
Check for outliers using Cook's distance
}
\examples{
check_outliers(mtcars, mpg, c(wt, hp, qsec))
}
