% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_vif.R
\name{check_vif}
\alias{check_vif}
\title{Check multicollinearity using VIF}
\usage{
check_vif(data, response, predictors)
}
\arguments{
\item{data}{A data frame.}

\item{response}{Unquoted name of the response variable.}

\item{predictors}{A call to `c()` with unquoted predictor variable names.}
}
\value{
A ggplot2 bar plot showing VIFs.
}
\description{
Check multicollinearity using VIF
}
\examples{
check_vif(mtcars, mpg, c(wt, hp, qsec))
}
