% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{diagnostic}
\alias{diagnostic}
\title{Run Diagnostics}
\usage{
diagnostic(x, ID = "ID", alpha = 0.05, group = "obs", graphs = T,
  influence = T, multicol = T, power = F, group.me = NULL)
}
\description{
Run Diagnostics
}

