% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{get.table}
\alias{get.table}
\title{Report summary statistics for different models}
\usage{
get.table(model, analysis = NULL, type = "summary", level = NULL,
  pca = 1)
}
\description{
Report summary statistics for different models
}

