% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{get_table}
\alias{get_table}
\title{Report summary statistics for different models}
\usage{
get_table(model, analysis = NULL, type = "summary", level = NULL,
  pca = 1)
}
\description{
Grab data from differnet model types
}
