% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_data.R
\docType{data}
\name{hist}
\alias{hist}
\title{Historical Data}
\format{An object of class \code{data.frame} with 10000 rows and 9 columns.}
\usage{
data(hist)
}
\description{
10000 training cases for the Diagnostics project.
The variables here are
Pn (Pneumonia), Te (Temperature), VTB (Visited a TB location), TB (Tuberculosis), Sm (Smoker), LC (Lung Cancer), Br (Bronchitis),
XR (X-Ray result), Dy (Dyspnea, or shortness of breath).
}
\examples{
data(hist)
}
\keyword{datasets}
