% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{kappamer}
\alias{kappamer}
\title{Check Kappa}
\usage{
kappamer(fit, scale = TRUE, center = FALSE, add.intercept = TRUE,
  exact = FALSE)
}
\description{
Condition number test less than 10 or 20 (less than 30)
}

