% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{overdisp}
\alias{overdisp}
\title{Overdispersion for glmer}
\usage{
overdisp(model)
}
\description{
Approximate test of overdispersion from:
http://bbolker.github.io/mixedmodels-misc/glmmFAQ.html#overdispersion
}
\author{
Ben Bolker
}
