% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{trans.plot}
\alias{trans.plot}
\title{Plot possible transformations}
\usage{
trans.plot(model, t = c("log", "inverse", "^1/2", "^2", "^3"), boxcox = NA,
  find.best = FALSE, verbose = FALSE)
}
\description{
Plot possible transformations
}

