% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{trans}
\alias{trans}
\title{Transformations}
\usage{
trans(x, trans, boxcox = NA, anchor = TRUE, centre = FALSE)
}
\description{
Function to easily transform variables in a variety of ways.
By default, variables are anchored at one to maximize the effectiveness of the transformations.
}
