% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{uni_kappa}
\alias{uni_kappa}
\title{Universal calculation of Condition Number (kappa)}
\usage{
uni_kappa(model, scale = FALSE, center = FALSE, exact = FALSE,
  add_intercept = TRUE)
}
\description{
Calculates Condition Number of matrix for common model types
}
\details{
For non-mixed models (lm and glm) uses kappa() from base package. For mixed models
uses function from Austin F. Frank:
https://github.com/aufrank/R-hacks/blob/master/mer-utils.R

Condition numbers should be less than 10 or 20 or you have multicollinearity
"values greater than 30 indicate collinearities that require attention." (Quinn and Keough
2002)

Note that kappa(lm) == uni_kappa(lm) but kappa(glm) != uni_kappa(glm)
Has to do with method of calculation, use with caution
}
