% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{uni_vif}
\alias{uni_vif}
\title{Universal calculation of Variance Inflation Factors (VIF)}
\usage{
uni_vif(model)
}
\description{
Calculates VIF for most models
}
\details{
For non-mixed models (lm and glm) uses vif from car package. For mixed models
uses function from Austin F. Frank:
https://github.com/aufrank/R-hacks/blob/master/mer-utils.R

"VIF values greater than ten suggest strong collinearity." (Quinn & Keough, 2002)
}
