% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checking.R
\name{diagnostic}
\alias{diagnostic}
\title{Run Diagnostics}
\usage{
diagnostic(model, group = "obs", graphs = TRUE, influence = TRUE,
  multicol = TRUE, verbose = FALSE)
}
\description{
Run complete diagnostics
}

